bug:[CORE-942]syncing EE and CC code backend apis, admin module

+++ api/pacman-api-admin/src/main/java/com/tmobile/pacman/api/admin/repository/service/AzureAccountServiceImpl.java
@@ -15,8 +15,11 @@ import com.microsoft.azure.management.Azure.Authenticated;
 import com.microsoft.azure.management.resources.Subscription;
 import com.tmobile.pacman.api.admin.domain.AccountValidationResponse;
 import com.tmobile.pacman.api.admin.domain.CreateAccountRequest;
+import com.tmobile.pacman.api.admin.repository.AzureAccountRepository;
+import com.tmobile.pacman.api.admin.repository.model.AccountDetails;
 import com.tmobile.pacman.api.commons.Constants;
 import com.tmobile.pacman.api.commons.config.CredentialProvider;
+import com.tmobile.pacman.api.commons.repo.PacmanRdsRepository;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,13 +27,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 @Service
 public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implements AccountsService{
-
 
     private static final Logger LOGGER= LoggerFactory.getLogger(AzureAccountServiceImpl.class);
     public static final String MISSING_MANDATORY_PARAMETER = "Missing mandatory parameter: ";
@@ -38,12 +38,16 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
     public static final String FAILURE = "FAILURE";
     public static final String SUCCESS = "SUCCESS";
     public static final String SECRET_ALREADY_EXIST_FOR_ACCOUNT = "Secret already exist for account";
-
+    public static final String AZURE_ENABLED = "azure.enabled";
+    @Autowired
+    AzureAccountRepository azureAccountRepository;
 
     @Value("${secret.manager.path}")
     private String secretManagerPrefix;
     @Autowired
     CredentialProvider credentialProvider;
+    @Autowired
+    PacmanRdsRepository pacmanRdsRepository;
 
     @Override
     public String serviceType() {
@@ -53,10 +57,7 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
     @Override
     public AccountValidationResponse validate(CreateAccountRequest accountData) {
         LOGGER.info("Inside validate method of AzureAccountServiceImpl");
-
         AccountValidationResponse response=new AccountValidationResponse();
-        response.setTenantId(accountData.getTenantId());
-        response.setType(Constants.AZURE);
         String tenant=accountData.getTenantId();
         String secretData=accountData.getTenantSecretData();
         if(StringUtils.isEmpty(tenant) || StringUtils.isEmpty(secretData)){
@@ -67,24 +68,9 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
             return response;
         }
         try {
-            Map<String,Map<String,String>> credsMap = new HashMap<>();
-            Arrays.asList(secretData.split("##")).stream().forEach(cred-> {
-                Map<String,String> credInfoMap = new HashMap<>();
-                Arrays.asList(cred.split(",")).stream().forEach(str-> credInfoMap.put(str.split(":")[0],str.split(":")[1]));
-                credsMap.put(credInfoMap.get("tenant"), credInfoMap);
-            });
-
-            String clientId = credsMap.get(tenant).get("clientId");
-            String secret = credsMap.get(tenant).get("secretId");
-
-            ApplicationTokenCredentials applicationCreds = new ApplicationTokenCredentials(clientId, tenant, secret, AzureEnvironment.AZURE);
-            LOGGER.debug("Application creds:{}", applicationCreds);
-
-            Authenticated azureAuthenticated = Azure.authenticate(applicationCreds);
-
-            LOGGER.debug("Azure authenticated :{}", azureAuthenticated);
-            PagedList<Subscription> subscriptions = azureAuthenticated.subscriptions().list();
-            LOGGER.debug("Subscriptions fetched :{}", subscriptions);
+            response.setTenantId(accountData.getTenantId());
+            response.setType(Constants.AZURE);
+            authenticateAzureAccount(accountData);
             response.setValidationStatus(SUCCESS);
             response.setMessage("Azure account validation successful");
         }catch (Exception e){
@@ -96,6 +82,30 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
         return response;
     }
 
+    private static PagedList<Subscription> authenticateAzureAccount(CreateAccountRequest accountData) {
+        String tenant=accountData.getTenantId();
+        String secretData=accountData.getTenantSecretData();
+        Map<String,Map<String,String>> credsMap = new HashMap<>();
+        Arrays.asList(secretData.split("##")).stream().forEach(cred-> {
+            Map<String,String> credInfoMap = new HashMap<>();
+            Arrays.asList(cred.split(",")).stream().forEach(str-> credInfoMap.put(str.split(":")[0],str.split(":")[1]));
+            credsMap.put(credInfoMap.get("tenant"), credInfoMap);
+        });
+
+        String clientId = credsMap.get(tenant).get("clientId");
+        String secret = credsMap.get(tenant).get("secretId");
+
+        ApplicationTokenCredentials applicationCreds = new ApplicationTokenCredentials(clientId, tenant, secret, AzureEnvironment.AZURE);
+        LOGGER.debug("Application creds:{}", applicationCreds);
+
+        Authenticated azureAuthenticated = Azure.authenticate(applicationCreds);
+
+        LOGGER.debug("Azure authenticated :{}", azureAuthenticated);
+        PagedList<Subscription> subscriptions = azureAuthenticated.subscriptions().list();
+        LOGGER.debug("Subscriptions fetched :{}", subscriptions);
+        return subscriptions;
+    }
+
     @Override
     public AccountValidationResponse addAccount(CreateAccountRequest accountData) {
         LOGGER.info("Inside addAccount method of AzureAccountServiceImpl");
@@ -105,8 +115,8 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
             return validateResponse;
         }
         String tenantId = accountData.getTenantId();
-        String tenantName=accountData.getTenantName();
-        validateResponse=createAccountInDb(tenantId,tenantName, Constants.AZURE);
+        validateResponse=createAccountInDb(tenantId,tenantId, Constants.AZURE,accountData.getCreatedBy());
+
         if(validateResponse.getValidationStatus().equalsIgnoreCase(FAILURE)){
             LOGGER.info("Account already exists");
             return validateResponse;
@@ -126,13 +136,14 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
             CreateSecretResult createResponse = secretClient.createSecret(createRequest);
             LOGGER.info("Create secret response: {}", createResponse);
             //update azure enable flag for scheduler job
-            String key="azure.enabled";
-            String value = "true";
-            String application = "job-scheduler";
-            updateConfigProperty(key, value, application);
+            updateConfigProperty(AZURE_ENABLED, TRUE, JOB_SCHEDULER);
             validateResponse.setValidationStatus(SUCCESS);
             validateResponse.setMessage("Account added successfully. Tenant id: "+tenantId);
-
+            PagedList<Subscription> subscriptions = authenticateAzureAccount(accountData);
+            String query="INSERT IGNORE INTO cf_AzureTenantSubscription (tenant,subscription,subscriptionName) VALUES (?,?,?)";
+            for(Subscription subscription : subscriptions){
+                pacmanRdsRepository.update(query,subscription.tenantId(), subscription.subscriptionId(),subscription.displayName());
+            }
         }catch (ResourceExistsException e){
             LOGGER.error(SECRET_ALREADY_EXIST_FOR_ACCOUNT);
             validateResponse.setValidationStatus(FAILURE);
@@ -155,14 +166,10 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
         AccountValidationResponse response=new AccountValidationResponse();
         StringBuilder validationErrorDetails=new StringBuilder();
         String tenantId=accountData.getTenantId();
-        String tenantName= accountData.getTenantName();
         String tenantSecretData=accountData.getTenantSecretData();
 
         if(StringUtils.isEmpty(tenantId)){
             validationErrorDetails.append(MISSING_MANDATORY_PARAMETER +" Tenant Id\n");
-        }
-        if(StringUtils.isEmpty(tenantName)){
-            validationErrorDetails.append(MISSING_MANDATORY_PARAMETER +" Tenant Name\n");
         }
         if(tenantSecretData==null){
             validationErrorDetails.append(MISSING_MANDATORY_PARAMETER +" Tenant Secret Data\n");
@@ -180,30 +187,47 @@ public class AzureAccountServiceImpl extends AbstractAccountServiceImpl implemen
     }
 
     @Override
-    public AccountValidationResponse deleteAccount(String tenantId) {
-        LOGGER.info("Inside deleteAccount method of AzureAccountServiceImpl. Tenant id: {}",tenantId);
+    public AccountValidationResponse deleteAccount(String subscription) {
+        LOGGER.info("Inside deleteAccount method of AzureAccountServiceImpl. Subscription id: {}",subscription);
         AccountValidationResponse response=new AccountValidationResponse();
         try{
             //find and delete cred file for account
             response.setType(Constants.AZURE);
-            response.setTenantId(tenantId);
             response.setValidationStatus(SUCCESS);
             response.setMessage("Account deleted successfully");
-            BasicSessionCredentials credentials = credentialProvider.getBaseAccCredentials();
-            String region = System.getenv("REGION");
-            String roleName= System.getenv(PALADINCLOUD_RO);
-            AWSSecretsManager secretClient = AWSSecretsManagerClientBuilder
-                    .standard()
-                    .withCredentials(new AWSStaticCredentialsProvider(credentials))
-                    .withRegion(region).build();
-            String secretId=secretManagerPrefix+"/"+roleName+"/azure/"+tenantId;
+            List<String> tenantId=azureAccountRepository.findTenantBySubscription(subscription);
+            azureAccountRepository.deleteById(subscription);
+            if(tenantId.isEmpty())
+            {
+                response.setValidationStatus(FAILURE);
+                response.setErrorDetails("Tenant id not found");
+                response.setMessage("Account deletion failed");
+                return response;
+            }
+            response.setTenantId(tenantId.get(0));
+            List<String> subscriptions=azureAccountRepository.findSubscriptionByTenant(tenantId.get(0));
+            if(subscriptions.isEmpty()) {
+                BasicSessionCredentials credentials = credentialProvider.getBaseAccCredentials();
+                String region = System.getenv("REGION");
+                String roleName = System.getenv(PALADINCLOUD_RO);
+                AWSSecretsManager secretClient = AWSSecretsManagerClientBuilder
+                        .standard()
+                        .withCredentials(new AWSStaticCredentialsProvider(credentials))
+                        .withRegion(region).build();
+                String secretId = secretManagerPrefix + "/" + roleName + "/azure/" + tenantId.get(0);
 
-            DeleteSecretRequest deleteRequest=new DeleteSecretRequest().withSecretId(secretId).withForceDeleteWithoutRecovery(true);
-            DeleteSecretResult deleteResponse = secretClient.deleteSecret(deleteRequest);
-            LOGGER.info("Delete secret response: {} ",deleteResponse);
+                DeleteSecretRequest deleteRequest = new DeleteSecretRequest().withSecretId(secretId).withForceDeleteWithoutRecovery(true);
+                DeleteSecretResult deleteResponse = secretClient.deleteSecret(deleteRequest);
+                LOGGER.info("Delete secret response: {} ", deleteResponse);
 
-            //delete entry from db
-            response=deleteAccountFromDB(tenantId);
+                //delete entry from db
+                response = deleteAccountFromDB(tenantId.get(0));
+            }
+            List<AccountDetails> onlineAccounts=findOnlineAccounts(STATUS_CONFIGURED,Constants.AZURE);
+            if(onlineAccounts==null || onlineAccounts.isEmpty()){
+                LOGGER.debug("Last account for Azure is deleted, disabling aws enable flag");
+                updateConfigProperty(AZURE_ENABLED,FALSE,JOB_SCHEDULER);
+            }
             response.setType(Constants.AZURE);
         } catch (SdkClientException e) {
             LOGGER.error("Error in deleting the creds file json: {}", e.getMessage());
